requirements.txt
streamlit>=1.28.0
google-genai>=0.7.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0
